contact

+++ build.sh
@@ -1,9 +1,9 @@
 #!/bin/sh
-PAGES=index 
-# radien-index hr contact
+PAGES="index contact"
+# radien-index hr 
 for page in $PAGES
 do
-	mv -f $page.html $page.html-
+	[ -f $page.html ] && mv -f $page.html $page.html-
 	m4 $page.html.m4 > $page.html
 	ls -l $page.html
 done
